Change covert channel type, almost finished
We have exploited the TC flag in DNS queries. It works correctly,
run `make receive` in receiver and `make send` in sender, and after
both are finished, run `make compare`. It works kind of slowly,
and the documentation part is still not finished. We get a warning
while running the sender code, does not seem to affect the outcome.
Need someone else to check the code.